moved builder in to its own namespace
