+++ Bigtable/src/Filter.php
@@ -19,16 +19,16 @@ namespace Google\Cloud\Bigtable;
 
 use Google\Cloud\Bigtable\Filter\ChainFilter;
 use Google\Cloud\Bigtable\Filter\ConditionFilter;
-use Google\Cloud\Bigtable\Filter\FamilyFilter;
+use Google\Cloud\Bigtable\Filter\Builder\FamilyFilter;
 use Google\Cloud\Bigtable\Filter\FilterInterface;
 use Google\Cloud\Bigtable\Filter\InterleaveFilter;
-use Google\Cloud\Bigtable\Filter\KeyFilter;
-use Google\Cloud\Bigtable\Filter\LimitFilter;
-use Google\Cloud\Bigtable\Filter\OffsetFilter;
-use Google\Cloud\Bigtable\Filter\QualifierFilter;
+use Google\Cloud\Bigtable\Filter\Builder\KeyFilter;
+use Google\Cloud\Bigtable\Filter\Builder\LimitFilter;
+use Google\Cloud\Bigtable\Filter\Builder\OffsetFilter;
+use Google\Cloud\Bigtable\Filter\Builder\QualifierFilter;
 use Google\Cloud\Bigtable\Filter\SimpleFilter;
-use Google\Cloud\Bigtable\Filter\TimestampFilter;
-use Google\Cloud\Bigtable\Filter\ValueFilter;
+use Google\Cloud\Bigtable\Filter\Builder\TimestampFilter;
+use Google\Cloud\Bigtable\Filter\Builder\ValueFilter;
 use Google\Cloud\Bigtable\V2\RowFilter;
 
 /**
