+++ Bigtable/src/Filter/Builder/QualifierFilter.php
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-namespace Google\Cloud\Bigtable\Filter;
+namespace Google\Cloud\Bigtable\Filter\Builder;
 
-use Google\Cloud\Bigtable\Filter;
+use Google\Cloud\Bigtable\Filter\QualifierRangeFilter;
+use Google\Cloud\Bigtable\Filter\SimpleFilter;
 use Google\Cloud\Bigtable\V2\RowFilter;
 
 /**
