+++ Bigtable/tests/Unit/Filter/Builder/QualifierFilterTest.php
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-namespace Google\Cloud\Bigtable\Tests\Unit\Filter;
+namespace Google\Cloud\Bigtable\Tests\Unit\Filter\Builder;
 
+use Google\Cloud\Bigtable\Filter\Builder\QualifierFilter;
+use Google\Cloud\Bigtable\Filter\QualifierRangeFilter;
 use Google\Cloud\Bigtable\Filter\SimpleFilter;
-use Google\Cloud\Bigtable\Filter\ValueFilter;
-use Google\Cloud\Bigtable\Filter\ValueRangeFilter;
 use Google\Cloud\Bigtable\V2\RowFilter;
 use PHPUnit\Framework\TestCase;
 
@@ -27,43 +27,38 @@ use PHPUnit\Framework\TestCase;
  * @group bigtable
  * @group bigtabledata
  */
-class ValueFilterTest extends TestCase
+class QualifierFilterTest extends TestCase
 {
-    private $valueFilter;
+    private $qualifierFilter;
 
     public function setUp()
     {
-        $this->valueFilter = new ValueFilter;
+        $this->qualifierFilter = new QualifierFilter;
     }
 
     public function testRegex()
     {
-        $filter = $this->valueFilter->regex('v1');
+        $filter = $this->qualifierFilter->regex('v1');
         $this->assertInstanceOf(SimpleFilter::class, $filter);
         $rowFilter = new RowFilter();
-        $rowFilter->setValueRegexFilter('v1');
+        $rowFilter->setColumnQualifierRegexFilter('v1');
         $this->assertEquals($rowFilter, $filter->toProto());
     }
 
     public function testExactMatch()
     {
-        $filter = $this->valueFilter->exactMatch('v1');
+        $filter = $this->qualifierFilter->exactMatch('v1');
         $this->assertInstanceOf(SimpleFilter::class, $filter);
         $rowFilter = new RowFilter();
-        $rowFilter->setValueRegexFilter('v1');
+        $rowFilter->setColumnQualifierRegexFilter('v1');
         $this->assertEquals($rowFilter, $filter->toProto());
     }
 
-    public function testRange()
+    public function testRangeWithInFamily()
     {
-        $this->assertInstanceOf(ValueRangeFilter::class, $this->valueFilter->range());
-    }
-
-    public function testStrip()
-    {
-        $filter = $this->valueFilter->strip();
-        $rowFilter = new RowFilter();
-        $rowFilter->setStripValueTransformer(true);
-        $this->assertEquals($rowFilter, $filter->toProto());
+        $this->assertInstanceOf(
+            QualifierRangeFilter::class,
+            $this->qualifierFilter->rangeWithInFamily('cf1')
+        );
     }
 }
